update the build system : using bili
