+++ lib/README.md
@@ -1,8 +1,17 @@
 # Reactjs-popup
 
-[![Build Status](https://travis-ci.org/yjose/reactjs-popup.svg?branch=master)](https://travis-ci.org/yjose/reactjs-popup) [![BCH compliance](https://bettercodehub.com/edge/badge/yjose/reactjs-popup?branch=master)](https://bettercodehub.com/) [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![Build Status][build-badge]][build]
+[![version][version-badge]][package]
+[![downloads][downloads-badge]][npmtrends]
+[![MIT License][license-badge]][license]
 
-[![NPM](https://nodei.co/npm/reactjs-popup.png?compact=true)](https://nodei.co/npm/reactjs-popup/)
+[![All Contributors](https://img.shields.io/badge/all_contributors-2-orange.svg?style=flat-square)](#contributors)
+[![PRs Welcome][prs-badge]][prs]
+[![Code of Conduct][coc-badge]][coc]
+
+[![Watch on GitHub][github-watch-badge]][github-watch]
+[![Star on GitHub][github-star-badge]][github-star]
+[![Tweet][twitter-badge]][twitter]
 
 > built with react Fragment : react 16 Tiny 3KB
 
@@ -69,39 +78,43 @@ export default () => (
 
 You can find more examples in the [reactjs-popup home page](https://react-popup.netlify.com)
 
-## Contributing
+# Contributing
 
-### Clone Repo
+## Clone Repo
 
 Fork and then clone the repo
 
     git clone git@github.com:your-username/reactjs-popup.git
 
-### Start Developing
+## Start Developing
 
 Install all npm scripts:
 
-```bash
-npm install
-#or
-yarn install
-```
+    npm install
+    or
+    yarn install
 
-we use storybook in this project.before starting, be sure to have some basic knowledge https://storybook.js.org/
+we use a simple package called parcel-story it's a simple storybook alternative more info https://github.com/yjose/parcel-story
+
+Run parcel-story :
+
+```
+yarn start
+```
 
 Run Test in watch mode
 
-```bash
-npm run storybook
+```
+yarn test-watch
 ```
 
-To make contributing simply you need to create a new story under src/stories directory ( you can copy/past any story to start with ).
+To make contributing simply you need to create a new story with documentation under stories/src directory ( you can copy/past any story to start with ).
 
-In this story, you need to present the new features or the bug fix and don't forget to document your story.
+In this story, you need to present the new features or the bug fix and don't forget to comment your code :) .
 
 Make Changes. If you want to contribute check out the [help wanted](https://github.com/yjose/reactjs-popup/issues?q=is%3Aissue+is%3Aopen+label%3A%22help+wanted%22) issues for things that need fixing.
 
-Before submitting a pull request run `npm run test` to run the unit-tests.
+Before submitting a pull request run `npm run test` to run the unit-tests and `npm run eslint` to check for linting errors in your changes.
 
 ## Licensing
 
@@ -109,6 +122,26 @@ The code in this project is licensed under MIT license.
 
 # Show your support!
 
-That's all, thank you for your attention, please star the repo to show your support...
+That's all, thank you for your attention, please [![Star on GitHub][github-star-badge]][github-star] the repo to show your support...
 
-...we are all made of stars star !
+...we are all made of stars [![Star on GitHub][github-star-badge]][github-star] !
+
+[build-badge]: https://img.shields.io/travis/yjose/reactjs-popup.svg?style=flat-square
+[build]: https://travis-ci.org/yjose/reactjs-popup
+[version-badge]: https://img.shields.io/npm/v/reactjs-popup.svg?style=flat-square
+[package]: https://www.npmjs.com/package/reactjs-popup
+[downloads-badge]: https://img.shields.io/npm/dm/reactjs-popup.svg?style=flat-square
+[npmtrends]: http://www.npmtrends.com/reactjs-popup
+[license-badge]: https://img.shields.io/npm/l/reactjs-popup.svg?style=flat-square
+[license]: https://github.com/yjose/reactjs-popup/blob/master/LICENSE
+[prs-badge]: https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square
+[prs]: http://makeapullrequest.com
+[coc-badge]: https://img.shields.io/badge/code%20of-conduct-ff69b4.svg?style=flat-square
+[coc]: https://github.com/yjose/reactjs-popup/blob/master/CODE_OF_CONDUCT.md
+[github-watch-badge]: https://img.shields.io/github/watchers/yjose/reactjs-popup.svg?style=social
+[github-watch]: https://github.com/yjose/reactjs-popup/watchers
+[github-star-badge]: https://img.shields.io/github/stars/yjose/reactjs-popup.svg?style=social
+[github-star]: https://github.com/yjose/reactjs-popup/stargazers
+[twitter]: https://twitter.com/intent/tweet?text=Check%20out%20reactjs-popup%20by%20%40ElaziziYoussouf%20https%3A%2F%2Fgithub.com%2Fyjose%2Freactjs-popup%20%F0%9F%91%8D
+[twitter-badge]: https://img.shields.io/twitter/url/https/github.com/yjose/reactjs-popup.svg?style=social
+[all-contributors]: https://github.com/yjose/all-contributors
